+++ Turbulence/Turbulence/SQLInterface/workers/GetMHDWorker.cs
@@ -125,12 +125,12 @@ namespace Turbulence.SQLInterface.workers
             }
         }
 
-        public override float[] GetResult(TurbulenceBlob blob, SQLUtility.InputRequest input)
+        public override double[] GetResult(TurbulenceBlob blob, SQLUtility.InputRequest input)
         {
             throw new NotImplementedException();
         }
 
-        public override float[] GetResult(TurbulenceBlob blob, SQLUtility.MHDInputRequest input)
+        public override double[] GetResult(TurbulenceBlob blob, SQLUtility.MHDInputRequest input)
         {
             float xp = input.x;
             float yp = input.y;
@@ -151,9 +151,9 @@ namespace Turbulence.SQLInterface.workers
         /// The Lagrangian evaluation function [LagInterpolation.EvaluateOpt] was moved
         /// into the function and some loop unrolling was performed.
         /// </remarks>
-        unsafe public float[] CalcLagInterpolation(TurbulenceBlob blob, float xp, float yp, float zp, ref double[] lagInt)
+        unsafe public double[] CalcLagInterpolation(TurbulenceBlob blob, float xp, float yp, float zp, ref double[] lagInt)
         {
-            float[] up = new float[3]; // Result value for the user
+            double[] up = new double[3]; // Result value for the user
 
             if (spatialInterp == TurbulenceOptions.SpatialInterpolation.None)
             {
@@ -260,9 +260,9 @@ namespace Turbulence.SQLInterface.workers
                             a2 += b2 * a;
                             a3 += b3 * a;
                         }
-                        up[0] = (float)a1;
-                        up[1] = (float)a2;
-                        up[2] = (float)a3;
+                        up[0] = a1;
+                        up[1] = a2;
+                        up[2] = a3;
                     }
                 }
             }
